+++ dashboard/app/models/concerns/user/database_authentication_overrides.rb
@@ -1,0 +1,23 @@
+module Devise::DatabaseAuthenticationOverrides
+  extend ActiveSupport::Concern
+
+  include Devise::Models::DatabaseAuthenticatable
+  include User::ProviderFlags
+
+  def update_without_password(params, *options)
+    if params[:races]
+      self.races = params[:races].join ','
+    end
+    params.delete(:races)
+    super
+  end
+
+  def update_with_password(params, *options)
+    if encrypted_password.blank?
+      params.delete(:current_password) # user does not have password so current password is irrelevant
+      update(params, *options)
+    else
+      super
+    end
+  end
+end
